Move all fixtures to conftest
